add dark mode to games
