+++ src/spice2x/util/utils.h
@@ -28,6 +28,9 @@ static const char HEX_LOOKUP_UPPERCASE[16] = {
 
 const char *inet_ntop(short af, const void *src, char *dst, DWORD size);
 
+// make a window's native title bar follow the Windows dark/light app theme
+void set_window_dark_titlebar(HWND hWnd);
+
 static inline bool string_begins_with(const std::string &s, const std::string &prefix) {
     return s.compare(0, prefix.size(), prefix) == 0;
 }
